+++ packages/gatsby-theme-docs/src/layouts/release-notes-list.js
@@ -53,7 +53,7 @@ const LayoutReleaseNotesList = (props) => {
               {props.children}
             </PageContentInset>
           </LayoutPageContent>
-          <LayoutPageReleaseNotesFilters scrollToTop={props.scrollToTop} />
+          <LayoutPageReleaseNotesFilters />
         </LayoutPage>
         <LayoutFooter />
       </LayoutMain>
@@ -72,7 +72,6 @@ LayoutReleaseNotesList.propTypes = {
     excludeFromSearchIndex: PropTypes.bool.isRequired,
   }).isRequired,
   children: PropTypes.node.isRequired,
-  scrollToTop: PropTypes.func,
 };
 
 export default LayoutReleaseNotesList;
