+++ packages/gatsby-theme-docs/src/utils/scroll-to-top.js
@@ -1,0 +1,8 @@
+function scrollToTop() {
+  const layoutPage = document.querySelector('#top');
+  layoutPage.scrollIntoView({
+    block: 'start',
+  });
+}
+
+export default scrollToTop;
