+++ packages/gatsby-theme-docs/src/utils/topic-is-checked.js
@@ -1,0 +1,11 @@
+function topicIsChecked(filterTopicsArray, topic) {
+  if (filterTopicsArray) {
+    return Array.isArray(filterTopicsArray)
+      ? filterTopicsArray.includes(topic)
+      : topic === filterTopicsArray;
+  }
+
+  return false;
+}
+
+export default topicIsChecked;
